More accurate comments
I know only WHY should be documented, but to keep the code self-contained a procedural code was preferred to the use of helper sub-procedures. However, this has the disadvantage of not being able to use descriptive function names to self-document the code. Still, I believe this was the correct choise.